Initialize websocket manager in app state
