+++ project/backend/basic_functions/crawlers/utils.py
@@ -4,7 +4,6 @@ from google.genai import types
 from project.backend.app.schemas.response import ProductAnalysisResult
 from project.backend.app.manage.settings import get_settings
 from project.backend.app.manage.resilience import with_llm_resilience
-from fastapi import WebSocket
 import httpx
 import uuid
 import asyncio
@@ -16,37 +15,6 @@ FAKE_HEADERS = {
     "Referer": "https://www.instagram.com/",
     "Accept": "image/avif,image/webp,image/apng,image/svg+xml,image/*,*/*;q=0.8",
 }
-
-class ConnectionManager:
-    def __init__(self):
-        # 유저 ID별로 '여러 개'의 활성 웹소켓 연결(배열)을 관리합니다. (Strict Mode 대응)
-        self.active_connections: dict[str, list[WebSocket]] = {}
-
-    async def connect(self, websocket: WebSocket, user_id: str):
-        await websocket.accept()
-        if user_id not in self.active_connections:
-            self.active_connections[user_id] = []
-        self.active_connections[user_id].append(websocket)
-
-    def disconnect(self, websocket: WebSocket, user_id: str):
-        # 해당 유저의 특정 웹소켓만 찾아 배열에서 안전하게 제거
-        if user_id in self.active_connections:
-            if websocket in self.active_connections[user_id]:
-                self.active_connections[user_id].remove(websocket)
-            # 남은 커넥션이 하나도 없으면 딕셔너리에서 키 삭제
-            if not self.active_connections[user_id]:
-                del self.active_connections[user_id]
-
-    async def broadcast_to_user(self, user_id: str, message: str):
-        if user_id in self.active_connections:
-            dead_sockets = []
-            for ws in self.active_connections[user_id]:
-                try:
-                    await ws.send_text(message)
-                except Exception:
-                    dead_sockets.append(ws)
-            for dead_ws in dead_sockets:
-                self.disconnect(dead_ws, user_id)
 
 settings = get_settings()
 api_key = settings.google_api_key
